Fix wrong bit-length data type used to update register

+++ src/libbase/k60/watchdog.cpp
@@ -116,7 +116,7 @@ void Watchdog::StartupInitialize()
 	Unlock();
 
 	// Disable but allow reconfig on startup
-	uint32_t reg_h = 0;
+	uint16_t reg_h = 0;
 	SET_BIT(reg_h, WDOG_STCTRLH_DISTESTWDOG_SHIFT);
 #if MK60DZ10 || MK60D10
 	SET_BIT(reg_h, WDOG_STCTRLH_STNDBYEN_SHIFT);
@@ -125,6 +125,7 @@ void Watchdog::StartupInitialize()
 	SET_BIT(reg_h, WDOG_STCTRLH_STOPEN_SHIFT);
 	SET_BIT(reg_h, WDOG_STCTRLH_ALLOWUPDATE_SHIFT);
 	SET_BIT(reg_h, WDOG_STCTRLH_CLKSRC_SHIFT);
+
 	WDOG->STCTRLH = reg_h;
 }
